requirements.txt
matplotlib==3.9.0
mpmath==1.3.0
numpy==1.26.4
pykan==0.0.5
pymoo==0.6.1.1
scikit-learn==1.5.0
torch==2.3.0
tqdm==4.66.4
